terminaison client

+++ src/main/java/sn/askanbiBank/dao/Idao.java
@@ -6,7 +6,7 @@ public interface Idao <T>{
 	public void save (T t);
 	public List<T> liste();
 	public void update (T t);
-	public void delete(Long id);
+	public void delete(int id);
 	public T getByID (int id);
 
 }
